fixing the jca tests
