+++ test/gov/aps/jca/test/BasicExample.java
@@ -30,6 +30,7 @@ import gov.aps.jca.event.MonitorListener;
 
 /**
  * Simple basic usage test.
+ * 
  * @author <a href="mailto:matej.sekoranjaATcosylab.com">Matej Sekoranja</a>
  * @version $id$
  */
@@ -38,41 +39,39 @@ public class BasicExample {
     /**
      * Implementation of get listener.
      */
-	private class GetListenerImpl implements GetListener
-	{
-	    private DBR value = null;
-	    private CAStatus status = null;
-	    
-		/**
-		 * @see gov.aps.jca.event.GetListener#getCompleted(gov.aps.jca.event#GetEvent)
-		 */
-		public synchronized void getCompleted(GetEvent ev) {
-		    status = ev.getStatus();
-		    value = ev.getDBR();
-		    
-		    // notify retrival
-		    this.notifyAll();
-		}
-		
+    private class GetListenerImpl implements GetListener {
+        private DBR value = null;
+        private CAStatus status = null;
+
+        /**
+         * @see gov.aps.jca.event.GetListener#getCompleted(gov.aps.jca.event#GetEvent)
+         */
+        public synchronized void getCompleted(GetEvent ev) {
+            status = ev.getStatus();
+            value = ev.getDBR();
+
+            // notify retrival
+            this.notifyAll();
+        }
+
         public CAStatus getStatus() {
             return status;
         }
-        
+
         public DBR getValue() {
             return value;
         }
-	}
+    }
 
-	/**
+    /**
      * Implementation of monitor listener.
      */
     private static class MonitorListenerImpl implements MonitorListener {
-        
+
         /**
          * @see gov.aps.jca.event.MonitorListener#monitorChanged(gov.aps.jca.event.MonitorEvent)
          */
-        public void monitorChanged(MonitorEvent event)
-        {
+        public void monitorChanged(MonitorEvent event) {
             // immediately print info
             if (event.getStatus() == CAStatus.NORMAL)
                 event.getDBR().printInfo();
@@ -80,169 +79,174 @@ public class BasicExample {
                 System.err.println("Monitor error: " + event.getStatus());
         }
     }
-    
+
     /**
      * JCA context.
      */
     private Context context = null;
-    
+
     /**
      * Initialize JCA context.
-     * @throws CAException	throws on any failure.
+     * 
+     * @throws CAException
+     *             throws on any failure.
      */
     private void initialize() throws CAException {
-        
-		// Get the JCALibrary instance.
-		JCALibrary jca = JCALibrary.getInstance();
 
-		// Create a context with default configuration values.
-		context = jca.createContext(JCALibrary.CHANNEL_ACCESS_JAVA);
+        // Get the JCALibrary instance.
+        JCALibrary jca = JCALibrary.getInstance();
 
-		// Display basic information about the context.
+        // Create a context with default configuration values.
+        context = jca.createContext(JCALibrary.CHANNEL_ACCESS_JAVA);
+
+        // Display basic information about the context.
         System.out.println(context.getVersion().getVersionString());
-        context.printInfo(); System.out.println();
+        context.printInfo();
+        System.out.println();
     }
 
     /**
      * Destroy JCA context.
      */
     private void destroy() {
-        
+
         try {
 
             // Destroy the context, check if never initialized.
             if (context != null)
                 context.destroy();
-            
+
         } catch (Throwable th) {
             th.printStackTrace();
         }
     }
-    
-	/**
-	 * @param channelName
-	 */
-	public void execute(String channelName) {
 
-		try {
-		    
-		    // initialize context
-		    initialize();
+    /**
+     * @param channelName
+     */
+    public void execute(String channelName) {
 
-			// Create the Channel to connect to the PV.
-			Channel channel = context.createChannel(channelName);
+        try {
 
-			// Send the request and wait 5.0 seconds for the channel to connect to the PV.
-			context.pendIO(5.0);
+            // initialize context
+            initialize();
 
-			// If we're here, then everything went fine.
-			// Display basic information about the channel.
-			channel.printInfo();
+            // Create the Channel to connect to the PV.
+            Channel channel = context.createChannel(channelName);
 
-            /********************************************************************/ 
-            /***************************** sync get *****************************/ 
+            // Send the request and wait 5.0 seconds for the channel to connect
+            // to the PV.
+            context.pendIO(5.0);
+
+            // If we're here, then everything went fine.
+            // Display basic information about the channel.
+            channel.printInfo();
+
+            /********************************************************************/
+            /*****************************
+             * sync get
+             *****************************/
             /********************************************************************/
 
             System.out.println("\n------------------------------------------------\n");
             System.out.println("Sync get:");
-            
-    		// get request w/o callbacks will wait until flush/pendIO is called
+
+            // get request w/o callbacks will wait until flush/pendIO is called
             // (channel default 'type' and 'count' is used)
-    		DBR dbr = channel.get();
-    		context.pendIO(3.0);
-    		dbr.printInfo();
+            DBR dbr = channel.get();
+            context.pendIO(3.0);
+            dbr.printInfo();
 
-    		System.out.println();
+            System.out.println();
 
-    		dbr = channel.get(DBRType.STRING, 1);
-    		context.pendIO(3.0);
-    		String[] value = ((STRING)dbr).getStringValue();
-    		System.out.println("Read string value: " + value[0]);
-    		
-            /********************************************************************/ 
-            /**************************** async get *****************************/ 
+            dbr = channel.get(DBRType.STRING, 1);
+            context.pendIO(3.0);
+            String[] value = ((STRING) dbr).getStringValue();
+            System.out.println("Read string value: " + value[0]);
+
             /********************************************************************/
-    		
+            /****************************
+             * async get
+             *****************************/
+            /********************************************************************/
+
             System.out.println("\n------------------------------------------------\n");
             System.out.println("Async get:");
 
             // get request w/ callbacks are always issued immediately
-    		// not related to pendIO at all, but require pend_event (to be flushed also)
-    		GetListenerImpl listener = new GetListenerImpl();
-    		channel.get(listener);
-    		synchronized (listener)
-    		{
-    			// flush & get event back
-    			context.flushIO();
-    			// wait for response...
-    			listener.wait(3000);
-    		}
-    		
-    		if (listener.getStatus() == CAStatus.NORMAL)
-    		    listener.getValue().printInfo();
-    		else
-    		    System.err.println("Get error: " + listener.getStatus());
+            // not related to pendIO at all, but require pend_event (to be
+            // flushed also)
+            GetListenerImpl listener = new GetListenerImpl();
+            channel.get(listener);
+            synchronized (listener) {
+                // flush & get event back
+                context.flushIO();
+                // wait for response...
+                listener.wait(3000);
+            }
 
+            if (listener.getStatus() == CAStatus.NORMAL)
+                listener.getValue().printInfo();
+            else
+                System.err.println("Get error: " + listener.getStatus());
 
-            /********************************************************************/ 
-            /***************************** Monitors *****************************/ 
-            /********************************************************************/ 
-            
+            /********************************************************************/
+            /***************************** Monitors *****************************/
+            /********************************************************************/
+
             System.out.println("\n------------------------------------------------\n");
             System.out.println("Monitors:");
             System.out.println();
 
             // Create a monitor
-            Monitor monitor = 
-                channel.addMonitor(Monitor.VALUE, new MonitorListenerImpl());
+            Monitor monitor = channel.addMonitor(Monitor.VALUE, new MonitorListenerImpl());
             context.flushIO();
 
             // Sleep for 10 seconds (monitors will be printed out).
             Thread.sleep(10000);
-            
+
             // Clear the monitor
             monitor.clear();
 
             System.out.println("\n------------------------------------------------");
-            
-            /********************************************************************/ 
-            /********************************************************************/ 
-            /********************************************************************/ 
+
+            /********************************************************************/
+            /********************************************************************/
+            /********************************************************************/
 
             // Disconnect the channel.
-			channel.destroy();
+            channel.destroy();
 
-			// Flush all pending requests...
+            // Flush all pending requests...
             context.flushIO();
 
-			System.out.println("Done.");
+            System.out.println("Done.");
 
-		} catch (Throwable th) {
-			th.printStackTrace();
-		}
-		finally {
-		    // always finalize
-		    destroy();
-		}
+        } catch (Throwable th) {
+            th.printStackTrace();
+        } finally {
+            // always finalize
+            destroy();
+        }
 
-	}
-	
-	
-	/**
-	 * Program entry point. 
-	 * @param args	command-line arguments
-	 */
-	public static void main(String[] args) {
+    }
 
-	    // check command-line arguments
-		if (args.length != 1) {
-			System.out.println(
-				"usage: java " + BasicExample.class.getName() + " <pvname>");
-			System.exit(1);
-		}
-		
-		// execute
-		new BasicExample().execute(args[0]);
-	}
-	
+    /**
+     * Program entry point.
+     * 
+     * @param args
+     *            command-line arguments
+     */
+    public static void main(String[] args) {
+
+        // check command-line arguments
+        if (args.length != 1) {
+            System.out.println("usage: java " + BasicExample.class.getName() + " <pvname>");
+            System.exit(1);
+        }
+
+        // execute
+        new BasicExample().execute(args[0]);
+    }
+
 }
